backend/requirements.txt
# Django and REST Framework
Django==4.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.0.0

# Database
psycopg2-binary==2.9.6

# Environment variables
python-decouple==3.8
python-dotenv==1.0.0

# HTTP requests (for external API calls)
requests==2.31.0

# WSGI server for production
gunicorn==20.1.0

# Other dependencies
Pillow==10.0.0
